Stabilize SQLAlchemy async test isolation
1. Make startup DB cleanup fixture loop-safe (conftest.py) by using a sync session fixture with asyncio.run(...)
2. Add shared auth_app_with_tx fixture that overrides get_async_connection with a transaction-bound connection per test
3. Truncate accounting tables at fixture start and always roll back at fixture end for clean per-test isolation
4. Migrate persistence/subscription route tests to use auth_app_with_tx and async HTTPX client
5. Configure db.py to use NullPool when TESTING=True to avoid cross-event-loop asyncpg connection reuse
6. Remove per-test ENGINE.dispose() from test_routes.py to prevent engine churn/flakiness
