+++ rctab/db.py
@@ -4,12 +4,18 @@ from typing import AsyncGenerator, Final
 
 from sqlalchemy.ext.asyncio import create_async_engine
 from sqlalchemy.ext.asyncio.engine import AsyncConnection
+from sqlalchemy.pool import NullPool
 
 from rctab.settings import get_settings
 
 SETTINGS: Final = get_settings()
 
-ENGINE: Final = create_async_engine(str(SETTINGS.postgres_dsn))
+ENGINE_KWARGS: dict[str, object] = {}
+if SETTINGS.testing:
+    # Avoid sharing asyncpg pooled connections across event loops in tests.
+    ENGINE_KWARGS["poolclass"] = NullPool
+
+ENGINE: Final = create_async_engine(str(SETTINGS.postgres_dsn), **ENGINE_KWARGS)
 
 
 async def get_async_connection() -> AsyncGenerator[AsyncConnection, None]:
