+++ tests/conftest.py
@@ -1,7 +1,8 @@
+import asyncio
 import datetime
 import subprocess
 from pathlib import Path
-from typing import Any, Callable, Dict, Tuple
+from typing import Any, AsyncGenerator, Callable, Dict, Tuple
 from uuid import UUID
 
 import jwt
@@ -10,6 +11,7 @@ from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 from fastapi import FastAPI, Request
 from jose.jws import sign
+from sqlalchemy.ext.asyncio.engine import AsyncConnection
 from sqlalchemy.sql.expression import text
 
 from rctab.crud.auth import (
@@ -19,35 +21,78 @@ from rctab.crud.auth import (
     user_authenticated,
     user_authenticated_no_error,
 )
-from rctab.db import ENGINE
+from rctab.db import ENGINE, get_async_connection
 from rctab.settings import Settings
 from tests.test_routes import constants
 
 
 @pytest.fixture(scope="session", autouse=True)
-async def clear_database_once() -> None:
+def clear_database_once() -> None:
     """Clear accounting tables once before the test session starts."""
-    async with ENGINE.begin() as conn:
-        await conn.execute(
-            text(
-                """
-                TRUNCATE TABLE
-                    accounting.status,
-                    accounting.usage,
-                    accounting.costmanagement,
-                    accounting.allocations,
-                    accounting.approvals,
-                    accounting.persistence,
-                    accounting.emails,
-                    accounting.cost_recovery,
-                    accounting.finance,
-                    accounting.finance_history,
-                    accounting.subscription,
-                    accounting.cost_recovery_log
-                RESTART IDENTITY CASCADE
-                """
+
+    async def _clear() -> None:
+        async with ENGINE.begin() as conn:
+            await conn.execute(
+                text(
+                    """
+                    TRUNCATE TABLE
+                        accounting.status,
+                        accounting.usage,
+                        accounting.costmanagement,
+                        accounting.allocations,
+                        accounting.approvals,
+                        accounting.persistence,
+                        accounting.emails,
+                        accounting.cost_recovery,
+                        accounting.finance,
+                        accounting.finance_history,
+                        accounting.subscription,
+                        accounting.cost_recovery_log
+                    RESTART IDENTITY CASCADE
+                    """
+                )
             )
+
+    asyncio.run(_clear())
+
+
+@pytest.fixture
+async def auth_app_with_tx(auth_app: FastAPI) -> AsyncGenerator[FastAPI, None]:
+    """Override DB dependency with one transaction-bound connection per test."""
+    conn = await ENGINE.connect()
+    trans = await conn.begin()
+    await conn.execute(
+        text(
+            """
+            TRUNCATE TABLE
+                accounting.status,
+                accounting.usage,
+                accounting.costmanagement,
+                accounting.allocations,
+                accounting.approvals,
+                accounting.persistence,
+                accounting.emails,
+                accounting.cost_recovery,
+                accounting.finance,
+                accounting.finance_history,
+                accounting.subscription,
+                accounting.cost_recovery_log
+            RESTART IDENTITY CASCADE
+            """
         )
+    )
+
+    async def _get_async_connection_override() -> AsyncGenerator[AsyncConnection, None]:
+        yield conn
+
+    auth_app.dependency_overrides[get_async_connection] = _get_async_connection_override
+    try:
+        yield auth_app
+    finally:
+        auth_app.dependency_overrides.pop(get_async_connection, None)
+        if trans.is_active:
+            await trans.rollback()
+        await conn.close()
 
 
 @pytest.fixture
